rename serialData to seriesData and create staticData for non-series data
